refactor PR#182

+++ install/stubs/seeds/update/SystemSettingsTableSeeder.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace EvolutionCMS\Installer\Update;
 
 use Illuminate\Database\Seeder;
@@ -12,8 +13,6 @@ class SystemSettingsTableSeeder extends Seeder
      */
     public function run()
     {
-        $value = 0;
-
         $res = \DB::table('system_settings')
             ->where('setting_name', '=', 'aliaslistingfolder')
             ->orWhere('setting_name', '=', 'full_aliaslisting')
@@ -24,26 +23,26 @@ class SystemSettingsTableSeeder extends Seeder
         foreach ($res as $row) {
             $values[$row->setting_name] = $row->setting_value;
         }
-
-        if (isset($values['full_aliaslisting']) && $values['full_aliaslisting'] == 1) {
-            $value = 1;
-
-            if (isset($values['aliaslistingfolder']) && $values['aliaslistingfolder'] == 1) {
+        if (isset($values['full_aliaslisting']) || isset($values['aliaslistingfolder'])) {
+            $value = 0;
+            if (isset($values['full_aliaslisting']) && $values['full_aliaslisting'] == 1) {
+                $value = 1;
+            } elseif (isset($values['aliaslistingfolder']) && $values['aliaslistingfolder'] == 1) {
                 $value = 2;
             }
-        }
 
-        $insertArray = [
-            ['setting_name' => 'alias_listing', 'setting_value' => $value],
-        ];
-        \DB::table('system_settings')->insert($insertArray);
+            $insertArray = [
+                ['setting_name' => 'alias_listing', 'setting_value' => $value],
+            ];
+            \DB::table('system_settings')->insert($insertArray);
 
-        $deleteArray = [
-            'aliaslistingfolder',
-            'full_aliaslisting',
-        ];
-        foreach ($deleteArray as $value) {
-            \DB::table('system_settings')->where('setting_name', $value)->delete();
+            $deleteArray = [
+                'aliaslistingfolder',
+                'full_aliaslisting',
+            ];
+            foreach ($deleteArray as $value) {
+                \DB::table('system_settings')->where('setting_name', $value)->delete();
+            }
         }
     }
 }
